Доработка структуры проекта
