+++ src/main.js
@@ -1,4 +1,4 @@
-import {generateWaypoints} from "./mock/route-waypoint-data-mock";
+import {generateWaypoints} from "./mock/mock-utils";
 import {createTripInfoContainerTemplate} from "./view/trip-info-container";
 import {createTripInfoRouteTemplate} from "./view/trip-info-route";
 import {createTripCostTemplate} from "./view/trip-cost";
@@ -9,7 +9,7 @@ import {createEventForm} from "./view/event-form";
 import {createWaypointTemplate} from "./view/waypoint";
 import {renderTemplate} from "./utils";
 
-const RENDER_EVENTS_COUNT = 20;
+// шапка
 
 const tripMainElement = document.querySelector(`.trip-main`);
 const tripEventsElement = document.querySelector(`.trip-events`);
@@ -24,15 +24,22 @@ renderTemplate(tripInfoContainer, createTripInfoRouteTemplate(), `beforeend`);
 renderTemplate(tripInfoContainer, createTripCostTemplate(), `beforeend`);
 renderTemplate(tripControlsFirstHeaderElement, createNavMenuTemplate(), `afterend`);
 renderTemplate(tripControlsSecondHeaderElement, createFilterTemplate(), `afterend`);
+
+// сортировка
+
 renderTemplate(tripEventsElement, createTripSortingTemplate(), `beforeend`);
+
+// форма
+
 renderTemplate(tripEventsElement, createEventForm(), `beforeend`);
 
 const tripEventsContainerElement = document.querySelector(`.trip-events__list`);
 
+const RENDER_EVENTS_COUNT = 20;
 const events = generateWaypoints(RENDER_EVENTS_COUNT);
 
 events.forEach(() => {
   renderTemplate(tripEventsContainerElement, createWaypointTemplate(events), `beforeend`);
 });
 
-
+// листнеры
