+++ src/mock/route-waypoint-data-mock.js
@@ -1,9 +1,7 @@
-import {getRandomInteger, getRandomDataArr, getEndDate, getDateDiffer, getRandomStartDate} from "../utils";
-
 export const eventTypes = [`Taxi`, `Bus`, `Train`, `Ship`, `Transport`,
   `Drive`, `Flight`, `Check-in`, `Sightseeing`, `Restaurant`];
 
-export const eventDestinations = [`Amsterdam`, `Chamonix`, `Geneva`, `Vladivostok`, `New York`, `London`, `Paris`, `Bruxelles`];
+export const eventDestinations = [`Amsterdam`, `Minsk`, `Geneva`, `Vladivostok`, `New York`, `London`, `Paris`, `Bruxelles`];
 
 export const infoDescriptions = [
   `Lorem ipsum dolor sit amet, consectetur adipiscing elit.`,
@@ -22,71 +20,48 @@ export const infoDescriptions = [
 
 export const offers = [
   {
+    id: `luggage`,
     title: `Add luggage`,
     price: 30,
   },
   {
+    id: `comfort`,
     title: `Switch to comfort class`,
     price: 100,
   },
   {
+    id: `meal`,
     title: `Add meal`,
     price: 15,
   },
   {
+    id: `seats`,
     title: `Choose seats`,
     price: 5,
   },
   {
+    id: `train`,
     title: `Travel by train`,
     price: 40,
-  }];
-
-const generateRandomDescription = () => {
-  const count = getRandomInteger(1, 5);
-  let items = ``;
-
-  for (let i = 0; i <= count; i++) {
-
-    items += infoDescriptions[i];
-  }
-
-  return items;
-};
-
-const generateRandomPhoto = () => {
-  const generatePhoto = () => {
-
-    return `http://picsum.photos/248/152?r=` + getRandomInteger(0, 999);
-  };
-
-  return new Array(getRandomInteger(1, 4)).fill().map(generatePhoto);
-};
-
-const getRouteWaypointData = () => {
-  const hasOffers = Math.random() > 0.5;
-  const startDate = getRandomStartDate();
-  const endDate = getEndDate(startDate);
-
-  return {
-    eventType: getRandomDataArr(eventTypes),
-    eventDestination: getRandomDataArr(eventDestinations),
-    eventOffers: hasOffers === true ? offers : ``,
-    destinationInfo: {
-      description: generateRandomDescription(),
-      photo: generateRandomPhoto(),
-    },
-    eventTime: {
-      start: startDate,
-      end: endDate,
-      differ: getDateDiffer(startDate, endDate),
-    },
-    isFavorite: Boolean(getRandomInteger(0, 1)),
-    price: getRandomInteger(10, 1000)
-  };
-};
-
-export const generateWaypoints = (count) => {
-  return new Array(count).fill(``).map(getRouteWaypointData);
-};
-
+  },
+  {
+    id: `taxi`,
+    title: `Order a taxi`,
+    price: 25,
+  },
+  {
+    id: `breakfast`,
+    title: `Add breakfast`,
+    price: 11,
+  },
+  {
+    id: `carrent`,
+    title: `Rent a car`,
+    price: 60,
+  },
+  {
+    id: `wi-fi`,
+    title: `Access to Wi-Fi`,
+    price: 3,
+  },
+];
